init commit add /ban /unban /setname /makeadmin /fakemakeadmin and other remove /su

+++ restart.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
-read PID < node.pid;
-kill $PID
-node app.js >> node.log &
+if [[ -f /run/user/1000/node.pid ]]; then
+ read PID < /run/user/1000/node.pid;
+ echo "Kill $PID"
+ kill $PID
+fi
+node app.js &>> node.log &
+#read PID < /run/user/1000/node.pid;
+#echo "Start witch $PID"
